refactor(jangar): split terminal worktree infrastructure
